+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - TIMESHEET: Do not hide button to generate several lines.
 - TIMESHEET: Typo in name of two actions, rename 'generale' into 'generate'.
 - PARTNER : Fix can't change the address if partner is an employee.
+- DEBT RECOVERY : Fix unable to run debt recovery accounting batch.
 
 ## [5.1.5] - 2019-03-30
 ## Features